redid combat engineering explosive logic
