+++ src/main/scala/net/psforever/objects/ExplosiveDeployable.scala
@@ -1,7 +1,7 @@
 // Copyright (c) 2018 PSForever
 package net.psforever.objects
 
-import akka.actor.{Actor, ActorContext, ActorRef, Props}
+import akka.actor.Actor
 import net.psforever.objects.ce._
 import net.psforever.objects.definition.DeployableDefinition
 import net.psforever.objects.definition.converter.SmallDeployableConverter
@@ -11,8 +11,6 @@ import net.psforever.objects.serverobject.affinity.FactionAffinity
 import net.psforever.objects.serverobject.PlanetSideServerObject
 import net.psforever.objects.serverobject.damage.{Damageable, DamageableEntity}
 import net.psforever.objects.serverobject.damage.Damageable.Target
-import net.psforever.objects.sourcing.{DeployableSource, PlayerSource, SourceEntry}
-import net.psforever.objects.vital.etc.TrippedMineReason
 import net.psforever.objects.vital.resolution.ResolutionCalculations.Output
 import net.psforever.objects.vital.{SimpleResolutions, Vitality}
 import net.psforever.objects.vital.interaction.{DamageInteraction, DamageResult}
@@ -23,6 +21,7 @@ import net.psforever.services.Service
 import net.psforever.services.avatar.{AvatarAction, AvatarServiceMessage}
 import net.psforever.services.local.{LocalAction, LocalServiceMessage}
 
+import scala.annotation.unused
 import scala.concurrent.duration._
 
 class ExplosiveDeployable(cdef: ExplosiveDeployableDefinition)
@@ -36,7 +35,7 @@ object ExplosiveDeployable {
   final case class TriggeredBy(obj: PlanetSideServerObject)
 }
 
-class ExplosiveDeployableDefinition(private val objectId: Int)
+abstract class ExplosiveDeployableDefinition(private val objectId: Int)
   extends DeployableDefinition(objectId) {
   Name = "explosive_deployable"
   DeployCategory = DeployableCategory.Mines
@@ -44,6 +43,8 @@ class ExplosiveDeployableDefinition(private val objectId: Int)
   Packet = new SmallDeployableConverter
 
   private var detonateOnJamming: Boolean = true
+
+  private var stability: Boolean = false
 
   var triggerRadius: Float = 0f
 
@@ -54,15 +55,11 @@ class ExplosiveDeployableDefinition(private val objectId: Int)
     DetonateOnJamming
   }
 
-  override def Initialize(obj: Deployable, context: ActorContext): Unit = {
-    obj.Actor =
-      context.actorOf(Props(classOf[MineDeployableControl], obj), PlanetSideServerObject.UniqueActorName(obj))
-  }
-}
+  def Stable: Boolean = stability
 
-object ExplosiveDeployableDefinition {
-  def apply(dtype: DeployedItem.Value): ExplosiveDeployableDefinition = {
-    new ExplosiveDeployableDefinition(dtype.id)
+  def Stable_=(stableState: Boolean): Boolean = {
+    stability = stableState
+    Stable
   }
 }
 
@@ -98,24 +95,26 @@ abstract class ExplosiveDeployableControl(mine: ExplosiveDeployable)
     }
   }
 
-  def Interaction(obj: Vitality with FactionAffinity, damage: Int, data: DamageInteraction): Boolean = {
-    !mine.Destroyed && (if (damage == 0 && data.cause.source.SympatheticExplosion) {
-      Damageable.CanDamageOrJammer(obj, damage = 1, data)
-    } else {
-      Damageable.CanDamageOrJammer(obj, damage, data)
-    })
-  }
-
   final def HandleDamage(target: ExplosiveDeployable, cause: DamageResult, damage: Int): Unit = {
     target.LogActivity(cause)
     if (CanDetonate(target, damage, cause.interaction)) {
-      ExplosiveDeployableControl.explodes(target, cause)
-      ExplosiveDeployableControl.DestructionAwareness(target, cause)
+      ExplosiveDeployableControl.doExplosion(target, cause)
     } else if (target.Health == 0) {
       ExplosiveDeployableControl.DestructionAwareness(target, cause)
     } else {
       ExplosiveDeployableControl.DamageAwareness(target, cause, damage)
     }
+  }
+
+  def Interaction(obj: Vitality with FactionAffinity, damage: Int, data: DamageInteraction): Boolean = {
+    val actualDamage: Int = if (!mine.Definition.Stable && data.cause.source.SympatheticExplosion) {
+      math.max(damage, 1)
+    } else {
+      damage
+    }
+    !mine.Destroyed &&
+      Damageable.adversarialOrHackableChecks(obj, data) &&
+      (CanDetonate(obj, actualDamage, data) || Damageable.CanDamage(obj, actualDamage, data))
   }
 
   /**
@@ -130,21 +129,22 @@ abstract class ExplosiveDeployableControl(mine: ExplosiveDeployable)
     * @return `true`, if the target can be affected;
     *        `false`, otherwise
     */
-  def CanDetonate(obj: Vitality with FactionAffinity, damage: Int, data: DamageInteraction): Boolean = {
-    !mine.Destroyed && (if (damage == 0 && data.cause.source.SympatheticExplosion) {
-      Damageable.CanDamageOrJammer(mine, damage = 1, data)
-    } else {
-      Damageable.CanDamageOrJammer(mine, damage, data)
-    })
-  }
-
-  def explode(target: ExplosiveDeployable, cause: DamageResult): Unit = {
-    ExplosiveDeployableControl.explodes(target, cause)
-    ExplosiveDeployableControl.DestructionAwareness(target, cause)
+  def CanDetonate(obj: Vitality with FactionAffinity, @unused damage: Int, data: DamageInteraction): Boolean = {
+    val sourceDef = data.cause.source
+    val mineDef = mine.Definition
+    val explodeFromSympathy: Boolean = sourceDef.SympatheticExplosion && !mineDef.Stable
+    val explodeFromJammer: Boolean = ExplosiveDeployableControl.CanJammer(mine, data)
+    !mine.Destroyed && (explodeFromSympathy || explodeFromJammer)
   }
 }
 
 object ExplosiveDeployableControl {
+  def CanJammer(mine: ExplosiveDeployable, data: DamageInteraction): Boolean = {
+    Damageable.adversarialOrHackableChecks(mine, data) &&
+      data.cause.source.AdditionalEffect &&
+      mine.Definition.DetonateOnJamming
+  }
+
   /**
     * na
     * @param target na
@@ -152,22 +152,23 @@ object ExplosiveDeployableControl {
     * @param damage na
     */
   def DamageAwareness(target: ExplosiveDeployable, cause: DamageResult, damage: Int): Unit = {
-    if (!target.Jammed && Damageable.CanJammer(target, cause.interaction)) {
-      if ( {
-        target.Jammed = cause.interaction.cause match {
-          case o: ProjectileReason =>
-            val radius = o.projectile.profile.DamageRadius
-            Vector3.DistanceSquared(cause.interaction.hitPos, cause.interaction.target.Position) < radius * radius
-          case _ =>
-            true
+    if (
+      !target.Jammed &&
+        CanJammer(target, cause.interaction) &&
+        {
+          target.Jammed = cause.interaction.cause match {
+            case o: ProjectileReason =>
+              val radius = o.projectile.profile.DamageRadius
+              Vector3.DistanceSquared(cause.interaction.hitPos, cause.interaction.target.Position) < radius * radius
+            case _ =>
+              true
+          }
         }
+    ) {
+      if (target.Definition.DetonateOnJamming) {
+        explodes(target, cause)
       }
-      ) {
-        if (target.Definition.DetonateOnJamming) {
-          explodes(target, cause)
-        }
-        DestructionAwareness(target, cause)
-      }
+      DestructionAwareness(target, cause)
     }
   }
 
@@ -188,6 +189,11 @@ object ExplosiveDeployableControl {
       Zone.explosionDamage(Some(cause)),
       ExplosiveDeployableControl.detectionForExplosiveSource(target)
     )
+  }
+
+  def doExplosion(target: ExplosiveDeployable, cause: DamageResult): Unit = {
+    explodes(target, cause)
+    DestructionAwareness(target, cause)
   }
 
   /**
@@ -270,114 +276,5 @@ object ExplosiveDeployableControl {
       Vector3.DistanceSquared(g1.center.asVector3, g2.center.asVector3),
       Vector3.DistanceSquared(point1, point2)
     ) <= maxDistance
-  }
-}
-
-class MineDeployableControl(mine: ExplosiveDeployable)
-  extends ExplosiveDeployableControl(mine) {
-
-  def receive: Receive =
-    commonMineBehavior
-      .orElse {
-        case ExplosiveDeployable.TriggeredBy(obj) =>
-          setTriggered(Some(obj), delay = 200)
-
-        case MineDeployableControl.Triggered() =>
-          explodes(testForTriggeringTarget(
-            mine,
-            mine.Definition.innateDamage.map { _.DamageRadius }.getOrElse(mine.Definition.triggerRadius)
-          ))
-
-        case _ => ()
-      }
-
-  override def finalizeDeployable(callback: ActorRef): Unit = {
-    super.finalizeDeployable(callback)
-    //initial triggering upon build
-    setTriggered(testForTriggeringTarget(mine, mine.Definition.triggerRadius), delay = 1000)
-  }
-
-  def testForTriggeringTarget(mine: ExplosiveDeployable, range: Float): Option[PlanetSideServerObject] = {
-    val position = mine.Position
-    val faction = mine.Faction
-    val range2 = range * range
-    val sector = mine.Zone.blockMap.sector(position, range)
-    (sector.livePlayerList ++ sector.vehicleList)
-      .find { thing => thing.Faction != faction && Vector3.DistanceSquared(thing.Position, position) < range2 }
-  }
-
-  def setTriggered(instigator: Option[PlanetSideServerObject], delay: Long): Unit = {
-    instigator
-      .collect {
-        case _ if isConstructed.contains(true) && setup.isCancelled =>
-          //re-use the setup timer here
-          import scala.concurrent.ExecutionContext.Implicits.global
-          setup = context.system.scheduler.scheduleOnce(delay milliseconds, self, MineDeployableControl.Triggered())
-      }
-  }
-
-  def explodes(instigator: Option[PlanetSideServerObject]): Unit = {
-    instigator
-      .collect {
-        case _ =>
-          //explosion
-          mine.Destroyed = true
-          ExplosiveDeployableControl.DamageAwareness(
-            mine,
-            DamageInteraction(
-              SourceEntry(mine),
-              MineDeployableControl.trippedMineReason(mine),
-              mine.Position
-            ).calculate()(mine),
-            damage = 0
-          )
-      }
-      .orElse {
-        //reset
-        setup = Default.Cancellable
-        None
-      }
-  }
-}
-
-object MineDeployableControl {
-  private case class Triggered()
-
-  def trippedMineReason(mine: ExplosiveDeployable): TrippedMineReason = {
-    lazy val deployableSource = DeployableSource(mine)
-    val zone = mine.Zone
-    val ownerName = mine.OwnerName
-    val blame = zone
-      .Players
-      .find(a => ownerName.contains(a.name))
-      .collect { a =>
-        val name = a.name
-        assignBlameToFrom(name, zone.LivePlayers)
-          .orElse(assignBlameToFrom(name, zone.Corpses))
-          .getOrElse {
-            val player = PlayerSource(name, mine.Faction, mine.Position) //might report minor inconsistencies, e.g., exo-suit type
-            player.copy(unique = player.unique.copy(charId = a.id), progress = a.scorecard.CurrentLife)
-          }
-      }
-      .getOrElse(deployableSource)
-    TrippedMineReason(deployableSource, blame)
-  }
-
-  /**
-   * Find a player with a given name from this list of possible players.
-   * If the player is seated, attach a shallow copy of the mounting information.
-   * @param name player name
-   * @param blameList possible players in which to find the player name
-   * @return discovered player as a reference, or `None` if not found
-   */
-  private def assignBlameToFrom(name: String, blameList: List[Player]): Option[SourceEntry] = {
-    blameList
-      .find(_.Name.equals(name))
-      .map { player =>
-        PlayerSource
-          .mountableAndSeat(player)
-          .map { case (mount, seat) => PlayerSource.inSeat(player, mount, seat) }
-          .getOrElse { PlayerSource(player) }
-      }
   }
 }
